Check `isok()` in a bunch of places where it was missing
This is not an exhaustive fix of all `getdir()`->missing `isok()`. There are a lot of `getdir()` calls! It is some.
Most of these didn't just crash right on their own, but with the right circumstances (like monsters in the top-left corner for some) would crash.

Fix RAISE_UNDEAD not checking isok() after getting direction

Check isok() in untrap()

Check `isok()` after getting a direction for chatting

Check `isok()` when using bullwhips
I was able to yoink a xorn from near the top-right corner of the map into the bottom-left corner where I was standing. Hilarious, but clearly not okay.

Sensorpacks check `isok()`
No message or charge used, same as cancelling the getdir prompt. Can't be bothered to come up with a better response (like a stethoscope's YAFM).

Ok I think this one was actually fine but lets be sure