Replace "Antimagic Shield" with "Abjuration"
No one used antimagic shield and the spell was kinda boring. Kick it to the curb and say hello to Abjuration!

It attempts to dispel all adjacent summoned monsters. If they fail their `resist()`, they vanish. If they succeed their `resist()`, if they were a temporary summon to begin with, their duration is reduced (permanent summons are fully immune when they resist).