Suppress `rloc()`'s impossible in a safe circumstance
A teleporting monster that already has a place on the map should not cause an `impossible()` when `rloc()` fails to find a new place for it. This can happen when the level is full, and the monster's current place is no longer acceptable (eg a boulder has been added to it). However, this is fine, since the monster remains in place.
Adjust behaviour appropriately: the monster that failed to teleport is still able to take other actions.