Numina doesn't grant Omnisense
Numina already grants `DETECT_MONSTERS`, which is level-wide monster *detection*. This is what it should be.

What `OMNISENSE` does is make infrared/blood/life/earth/all 'sense' cover the whole level, but these visiontypes are treated as 'sight' and not 'sense', which makes for awkward "I can see the foo, interrupt current action".