Godlist and Artilist *should* be headers
I was informed that `#include`ing `"../src/file.c"` was "really cursed". And they're right, those files should be headers.
I previously changed artilist to be a source file because we needed a single instance of artilist to be shared across all dnh files that use it -- well, dynamic artifact code did exactly that by mallocing it in, and the base artilist was only used to populate the malloc'd one. So now it can be a header again safely.
These headers should ONLY be included for `util/makedefs` (and other `/util`s) and in the single `.c` for each where the main one is being init/save/restore-d.

For godlist, this means that in the future stuff like godanger can be saved with the god structure with no problems. Exception: god names are constants, so those can't be modified in-game and saved (without changing the god structure first).