Full handling of non-pantheon gods
Altars dedicated to a god that isn't yours are effectively cross-aligned (ex water prayer gives unholy water) and can be converted to (for aligned gods).

If you are a follower of a non-pantheon god, that deity appears as a 4th entry in ctrl-x self inspection (very relevant for Binders).

Fix Binders' pantheon being out-of-whack in the Astral Plane.