Better materials give at least some protection.
All materials get a flat DR and flat AC rating. If a zero-defense item is sent to the material function it gets a value based on these flat values.

In practice, this means that leather items that are normally cloth get 1 DR, and dragonhide items get some AC and DR.