Fix the bonkers enough_dogs() code (thanks K2 :D )
Looks like a five-year-old case of late-night coding leading to unfinished code.

The current code has different results depending on the order that pets happen to be in on the fmon chain.
-a weaker dog with a higher tameness will be untamed if it is found first, and retained if it is found second.

The new code should keep the higher-level pet, or the more tame pet IF level is tied, or the highest-max-level pet if both are tied, or the highest-difficulty pet if all three are tied