Expand `deadmonster` to 2 bits to mark 'dead' as well as 'purge'.
DEADMONSTER() generally just needs the 1st bit, but we want the 2nd bit to be sure that we're only purging monsters once.

Set the 1st bit when we would previously be setting (or confirming) that monster hp <1 and isn't recovering.
Set the 2nd bit after unlinking.

Fixes issues of monsters dying because of `mongone()`/`mondead()`/etc before they reach `m_detach()`.