Aphanacts: Paths of Law insight monsters
Surveyors/Scouts/Spies from an ancient civilization. The records of what they've seen are inscribed on red-gold disks they cary with them.

Aphanactonan monsters have auto-hit attacks that deal damage based on the level of the target.

To read the disk records, the PC must have pierced the veil

Aphanactonan audient
-A non-humanoid, hovering microphone (can't see)
-Explodes on death, similar to a gas spore but more powerful.
-1 - 8 insight to see
Aphanactonan record
-Reading this grants the PC knowledge of 2 to 8 item types

Aphanacton assessor
-Humanoid observer (can't hear)
-Has a primary study gaze, paralysis counter-gaze, and beholder gaze
-20 - 36 insight to see
Aphanactonan archive
-Reading this grants
--Knowledge of 4 to 16 item types
--4 draws for:
---1-4 random spells
---1-8 random seals
---2-8 random wards
---1-2 bonus to-hit, damage, and AC
---note: spells, seals, and wards have diminishing returns because duplicates are possible.

Each Paths of Law level has 0 to 4 assessors and 4 to 8 audients. This may not be balanced or desireable, especially as the Paths get more interesting.