Make the Star of Hypernotus display correctly w.r.t. the "star sapphire" material
Now shows as an "engraved blue-green star-shaped stone" when unidentified, and if you don't know the "star sapphire" gem. If you do know the star sapphire gem, shows up as an "engraved blue-green star-shaped star sapphire".
Note: star sapphires are a kind of sapphire with a star-shaped pattern on/in them, so star-shaped star sapphire is not redundant (I think).
Previously, this would show as an "engraved blue-green blue-green star-shaped stone", or an "engraved star sapphire blue-green star-shaped stone" or something like that. This fixes that.

Does some shenanigans to do so - special-cases the star of hypernotus to not have a material descriptor ever, and makes the 'artifact description' into "blue-green star shaped %s". The code that formats artifact description "%s"s into base names is then also special-cased to choose between "stone" and "star sapphire" as a base name instead.