Bugfix/nasty hack: Handle gold scroll polymorph to gold piece in dodip.
The game is in an inconsistent state at this point (because there's a gold object loose in the player's inventory), so the inventory handling functions do the wrong thing by subtracting the quantity of gold from the player's count. :., increment the player's gold count by 2x and then call useup on the gold pile.