Add a hd_size() function that always returns 8.
In the future, it would be nice to have HD be size or monster type dependent or similar.

The existing code uses the magic number 8 everywhere, so first everywhere that "8" is mentioned will need to be audited and replaced with this function.

However, this is added as groundwork so that new occurances of HD size can use it.