Make GDCB's weight actually adjustable
This means any future changes to its weight independent of its base item type or material need to be handled in artifact_weight, which isn't ideal if it's going to ever change weights a lot. Should be able to use ovar1 or something to track its status though so hey.