Switch up how the Stranger's harrasment works
The Yellow harrasment timer only runs while the stranger is on the level but hidden by insight.

The harrasment timer is about 10x shorter (but still about 2x longer than the Wizards).

The timer runs 5x faster as a result of the stranger's gaze.

The timer starts at 555 turns and doesn't start running untill the PC claims the quest artifact (after which, again, it only runs while the Stranger is on the level).
-The timer always runs for non-Madmen, so if you encounter the Stranger in bones it will begin following and harrassing the PC.

An extra "spell like" harrasment effect has been added.
--Due to limitations in the monster spellcasting code, a hostile monster is selected to cast the spells at you.
--This monster prints a special message and gains a level.
--If no monster can see you, 5-55 turns are put back on the timer.
--Effects:
---Punishment-like (you don't get a reistance roll vs. this)
---Sickness
---Level drain
---Cause wounds
---Nightmare
---Finger of Death-like
---All hostile monsters get +5 encouragement
---Poison cloud
---Acid rain
---Fire
---Cold