Buff/fix release from care
Fix: it's supposed to behead. The vorpal change lost this.
-Buff: Vs. large creatures with no head it does the "cleaving vorpal" 2x damage effect.

Buff: It cancels the target when vorpal effect triggers, which means that Trolls shouldn't revive (and spellcasters become far weaker if they lifesave)
-I believe this will also short-circuit a number of intrinsic resurrection effects like illuminated.

Buff: It vorpals 10% of the time instead of 5% (vs. vorpal blade, which gets an extra die and whose dice explode).