The Chaos 1 Black Crystal can lifesave the PC up to twice
Once on its own, and once if the other four crystals are in open inventory.
(The 5-crystal version is used first, since it's less advantageous)

Once crystals have been used to life save, they crack. The 5-crystal version cracks the 4 elemental crystals, the 1-crystal version cracks the black crystal.
-The crystals are not destroyed, and retain all their properties (other than life saving) if cracked
-If the 1-crystal version is used first (cracking the black crystal) the 5-crystal version can still be used normally.

If found in bones, the crystals will be uncracked. If left in bones but converted to ordinary crystal balls during loading, the crystal balls will still be cracked (this is cosmetic only)

Revised-Chaos is troublsome enough to fight that I doubted that the Black Crystal was worth it (even with its spells and skills).