Allow specifying alternate dungeon branches in one .def file
By default, they are all weighted equally, but the `select_alternate()` function may be given special cases.
Branches are numbered starting from 0, in the order they are written in dungeon.def.

Note: Chainbranches off of an alternate must exist on all alternates, or else it is possible the root of the chainbranch won't have been generated and the dungone linker will crash. Let me know if this is ever a feature we'd want to have, I can probably make it happen :)