Adjust pokedex to allow it to display effective wards
Filter on wards you actually know.

Add it to the default pokedex setting

Finish refactoring hamsa and yellow sign

Adjust description of elder sign and elder eye to suggest checking the pokedex.

Could use another refactoring to reduce complexity of the pokedex checks and reduce liklihood of divergence in the future.