Outlands prison and throne room fixes.
Previously, if an outlands throne room was up against some trees it could spawn an empty cell in the trees. This was kind an eyesore and also too random.

This creates a separate "jail" building function and blocks cell placement for throne rooms in the outlands.

Jails have a row of 3x3 cells containing non-neutral outsiders and adventurers. The equipment is in a chest guarded by a cuprilach. The cells extend one square further north and south than the typical building, so they are cleared and checked. The cell-making code is properly compartmentalized (this whole section needs a refactoring >_<).

Moves the wallification step to the end of outlands feature generation.