Adjust mk_mplayer to accept makemon flags.
Use MM_GOODEQUIP instead of the old boolean value.

Don't second-guess the goodequip input.

Fixes the bug that probably lead to the above code (two places in sp_lev try to send in MM_ADJACENTOK in the spot meant for the boolean, when placing healer monsters in the madman quest filler rooms).