Refactor and slightly improve marilith weapon selection
Refactor check for valid weapon and weapon selection into their own functions.

Monsters use the last n weapons in their inventories, rather than the first n. Since monster inventories are first-in-last-out, the first n weapons are the random junk it's picked up of the ground while the last n are the weapons it spawned with or that the player gave it (I actually thought that monster inventories were first-in-first-out when implementing marilith arms >_<).

Monsters with marilith arms ignore offhand weight limits if the weapon is an ok marilith arm weapon. Monsters may have bizarly weak offhand arms otherwise.