Crystal skull fix: Compartmentalize update code, call where needed.
The update was only being done in one of the two desummon functions, and was ineffective if the skull was dropped due to the skull being on no chain at the point the update needed to happen.

The refactored update code allows for an object to be sent in or finding the skull on its own, and has various impossible() state dection cases.

Also updates more madnesses.