Reduce average insight gain from aphanactonan disks
Now only has a 1-in-existing insight chance to grant an extra insight point (archives always grant 1 insight, but may grant up to 8 more).

The law quest isn't such an overload of free insight now.

Also changes the order in which sanity loss happens with archives, so that the temp madness message will be the last thing printed.