Artifact weight/mat handling
Move handling out of artifact code and into normal weight code.

If an artifact has both a nonstandard material and a nonstandard weight, its base mat is the artifact mat. Otherwise, the base mat is the OC mat.

No mater what, if the silver flame or mammon changes the mat of the artifact, the artifacts weight will be changed. It's just a question of what the new weight will be relative to.