Monster insight/sanloss and template adjustments:
Templates can confer insight-ness. Only one batch of insight is awarded per PM, so if an insight-granting monster is given an insight-granting template no extra insight is awarded, but if a non-insight monster like a deminymph is given an insight template insight is awarded. However, a second deminymph with a different insight-granting template won't award a second batch of insight.

Ditto sanloss

Pseudonatural template confers 1.5x level and max HP, moly template confers max hp (and already confered 1.5x level

The tomb herd is resistant to slashing and piercing weapons, but vulnerable to blunt. It also causes san loss (but not insight)