Bugfix: Mounted crystal skull monster crash
obj_no_longer_held is called INSIDE place_object(), AFTER the top object on the square is checked. But when a crystal skull is dropped and the mount desummoned, the mount's items are placed on top of the list, and are leaked as a result.

This simply gets the top object after calling no_longer_held