Xans don't have wings and can't fly.
Xans in dnh may be *named* after mosquitos, but they are apparents small, dense (14 points of natural AC!), supernatural insectoids with sharp but non-blood-sucking proboscises.

This is a better way to resolve the conflict between flying xans and AD_LEGS not being effective on flying monsters, since xans are the only monster with this damage type.