Implement Antimagic rift and Catapsi vortex lightsaber gems
I thought it was dummied out, but Death Knights had sabers with them.

They were just supposed to have a special appearance and a chance (10% here) to cancel apropriate targets.
-Catapsi cancels AD_PSON casters, Antimagic cancels AD_SPEL
--This isn't quite the same as what the consumed items would affect, but seems appropriate.

Since cancel_monst rolls resistance, and the expected result of hitting something with a lightsaber is that it loses its special abilities on account of being dead, this should have basically no game effect.