Reorder second status line
This reorders the second status line to put delayed instadeaths first,
as in NetHack 3.6.  It also removes the "Gold" status condition added
in the previous commit and instead shows "Stone" if you are stoned or
golded, as the effects and cures are the same for both.

The new order is:
Stone Slime Suffct FoodPois Ill <hunger> <encumbrance> Invl Blind Stun
 Conf Hallu Panic Stmblng Stggrng Babble Scream Faint Lev Fly Ride
The old order was:
<hunger> Invl Conf FoodPois Ill Blind Stun Panic Stmblng Stggrng
 Babble Scream Faint Hallu Lev Fly Ride Stone Gold Slime Sufct
 <encumbrance>

(This exact set of status conditions is impossible, as some of these
are mutually incompatible.)