Add abbreviated names for status conditions
Adds two sets of abbreviated names for status conditions and
encumbrance.  These are used when the full names don't fit, to avoid
them getting cut off.  statushilites work without requiring new
configuration for the abbreviated names, curses is unaffected, and
dumplogs are unaffected unless the status line is wider than
MAXCO (previously this would cause a buffer overflow).

The buffer overflow can no longer happen with status conditions as
they now use the new Snprintf/Strncat/Strncpy macros.  It can still
happen if you somehow manage to overflow the status line before status
conditions, but this should be currently impossible.  The value of
MAXCO is also increased to 200 (as in NetHack 3.6), as it's possible
to have a status line longer than 160, and the trailing spaces after
hunger states are removed.