Pirate Quest Tweaks/Bugfixes
Use the "jewel" keyword handling introduced for the DroHea quest to ensure that some of the random gems are valuable instead of glass.

Remove stray extra chest

Remove stray "contained" keyword (replace with "random") that was causing the random piles of gold and gems to end up in the last chest declared.