Update data.base
"Needs encyc entry" is a set of words that's bothered me forever since I started peeking at monst.c, so I'm writing a commit about it. 

New entries for Adamach, the sheep and its dire kin, several of the Archons, the orcs of the ages of stars, and a few others, with a bit more alphabetical organization to tie it all together. This isn't final in the slightest, but I figure it's a good place to start if I'm gonna be a code contributor of any kind.