DroHea quest tweaks
QoL: Add a wand of striking to both of the first two treasure chests. You may need to break a lot of iron bars to do this quest.

Add a guaranteed anulo dancer plague victim (in case you want one for a pet. All the existing ones were 50/50 so might not spawn).

Tweak: Add a glass set of eilistran armor and an eilistraee sword to that treasure chest(/sack), and make the "oLS fixed.
-Mostly just for fun, but
1) the second source of flying may be useful
2) the DroHea PC can actually USE a glass one, because not metal.

Tweak: Add a treasure sack with some sources of levitation to it to the cavern floor level.
-Somlia's games revealed that the player might wish to withdraw but be unable to do so if flight was lost or never acquired (cleric's feather fall?). This creates a sub-standard option for escape.

Add a mendicant drider plague-victim to the Pen'a temple (once again in case you want one as a pet).

Add a saddle to the Pen'a treasure chest.

Add a second unicorn horn to the Pen'a treasure chest. Unicorn horns are just as important as MR for a beastmaster run.

Fix the Panakein Archon's starting robes.

Add a... treasure chest of stolen pirate loot(?) to the quest goal level.
-Idk about the flavor. It seemed like a good idea at the time. I guess the kuo toa stole it from a beach or something.
-The real purpose of this chest is the small silver khakkharas, which are a thing you could farm off kuo toa whips, and which you might want for reasons.

Add some mindlessified/cordyceped/plagued Ilmater minions to the goal level.
-One more MR option
-I think 3 more unicorn horns