Tweak: add an enchant armor scroll to elf mad chest
They need a bunch potentially to fix their IEA. They get the armor on the first level, if they reach the chest they can at least fix one.

In some games the RNG will have smiled and provided a shop keeper that can fix errosion, this is to even things out a bit in RNG-disfavored games.