Fix figurines being able to generate any monster ever made ever
Figurines were patched a while ago so you could get unusable figurines (trophies etc. I guess) but unfortunately not all checks for them called the right functions. This led to makemon being called instead of animate_figurine, so you could wish for a figurine of demogorgon -> stone to flesh it -> die instantly probably but like, it's a legit copy of demogorgon not a doppelganger. Statues were fine, and now so are figurines.