Add the Lancet of Longinus
If you wish for it, you can now get this as a derived form of the Lance of Longinus, similar rules to other "malleable" weapons. Technically, this isn't malleable though - it just matches both scalpels and tiny lightsabers (only tiny) when trying to apply the artifact, but will never use those otyps for floor artifact purposes.

This does not affect size or material - a scalpel will always be medium sized and silver, and a lightsaber will always be forced to be tiny (not medium) and silver, though lightsabers are normally silver anyway.

If the Lance is either of those types, it will also be renamed to the "Lancet". This is merely cosmetic.