actually make warning farlook useful :)
okay, so first off this is kinda an egregious level of copypaste, since this doubles up on a lot of messaging already in the same file for nero's pokedex work
the biggest issue with those is that its set up to take a monster and some booleans and shit and that's not what u want out of warnings
the function i wrote just takes a flag and a int to set which flag, set with some cheap defines (though i could probably make that more universal). it is not extended outside pager.c
it outputs the general species name for MA flags, uses symbol name from monappearances, and the rest are adjectives/phrases that felt appropriate that work grammatically with "monsters that XXXXX"
it also incorporates the warning things from paimon tenebrous etc. for non-standard warning masks, so that's nice, so in the future add those to WARN_OF_MON and then here i guess

fixes #600 lmao