add #style toggle for passive-safe attacks and refactor #style code
refactors #style code to better handle generic "fighting forms" not just lightsaber vs. breaking out early for mystic forms vs. whatever
now, it handles mystic, lightsaber, passive toggle, and has handling for an as-of-yet non-existent knight forms. further categories of forms WILL need to have handling added, but it should be pretty simple - some function doFooForm() that handles the window code or whatever, and the criteria under which to have that as a list option / the code for adding it as a list option to the #style menu itself

also removes the inability to use #style ls forms if you're not wielding a lightsaber. now you can do it if wielding, uswapwielding, and if you have trained any lightsaber skills before. this DOES remove the check for weapon skill since its not always true, so better messaging for that could be nice, but it shouldn't be a huge deal - since you still have to be wielding a saber, and it'll tell you if you're unskilled in shii-cho or something. dunno. open to suggestions for improvements here
shitty knight form defines are based on the ones in the etherpad with absolutely no handling for any of them to do anything, just what's needed to set/check them without impossibles. there is no way to get those skills so they never show up and shouldn't matter

also refactors form setting/checking code slightly. now, each batch of 32 forms is mutually exclusive - so FFORMs 0-31, 32-63, etc. this is so that you can have both a lightsaber AND a knight form active, but not two ls or two knight forms active. this cannot be extended to monk forms since they can all be individually toggled.

passive avoidance code uses the new u.uavoid_passives to determine if you're a coward or not, and refactors my previosu stuff into the new non_contact_attk(attk) macro. its only available in #style if your current attack routine (NOT by the book, so including star emp ring or black web or something) includes a non_contact_attk(attk) to prevent it from confusing people when not relevant. however, if you're currently avoiding unsafe attacks, you can always turn that off - so that if you take off the ring or unbind web or something you're not stuck making no attacks forever or until you regain the ability to turn it off