add "detached from purpose" flag
this is basically "mindless but allows muse", though they don't spawn with the misc items. they do know how to use and will use offensive, defensive, and misc (invis pots etc.) items.

the intent is to provide a "not intelligent, no brain, but either subsumed will or not really all there anymore". the intent is for constellation pets, and constellations are now marked as MT_DETACHED instead of MT_MINDLESS, and it's not used for anything else.

note that anything that wants to check mindless checks mindless_mon, so in lieu of going through every single instance, MT_DETACHED is just added there. for "mindless but not detached", check mindless_muse instead.