Add "descendant" option to start game with inheritance
This copies most of the actual inheritance code from Inheritor dummy stuff, but removes the Inheritor race stuff entirely.
Instead, there is now flags.descendant, which is another thing to click through on starting game that's a y/n for "get inheritance/don't". Can be set with the boolean option "descendant", so adding "!descendant" to your rcfile will hide any and all traces of this option.
In-game, the only effects are if you did choose to be a descendant, in which case you spend your first artiwish on game start to get an inheritable artifact. That artifact loses any conflicting race, role, or alignmnet flags, but can only be chosen if it shares alignment, isn't already your role/pantheon. Furthermore, the message for descendants on starting/loading game says "descendant", and the ctrl-x lists the artifact you got at game start.

Again, for non-descendants none of this matters and there are no gameplay changes, just that first menu pre-game start.

Also, I added a bunch more artifacts to the inheritable list - this definitely needs more tweaking before this is really a 'finished' list, but as a feature it works fine.

possibly todo?
- can't really default this to off, but this could always just be a forced rcfile option (now that I've already written the menu lmao) if the menu popup is undesireable for beginners
- i only wrote tty/curses menus since the others are outdated as fuck and i don't know/want to know how to build with them