Make Apotheosis Veil bypass most AC/DR on headbutt attacks
Since it protects from extradimensional attacks, it should be able to
attack extradimensionally itself.

This is checked in roll_mdr_detail for monsters and slot_udr for
players.  Since they weren't passed the attack, it was impossible to
tell if it was an AT_BUTT attack or not.  Now roll_udr,
roll_udr_detail, slot_udr, roll_mdr, and roll_mdr_detail are passed
aatyp, with AT_ANY being passed where it's not applicable.  I'm not
sure if this is the best solution - maybe they should be passed the
entire attack instead.