Show TimeStop/BlowingWinds duration in statusline
Currently tty-only.  Potentially this could cause the statusline to
exceed MAXCO, but that's unlikely to happen in practice (it requires
lots of status conditions and a high duration of
TimeStop/BlowingWinds) and should just result in the abbreviated
version being used instead.  Maybe this should also be done for
Misotheism/Catapsi/Deadmagic, but those usually have a much longer
duration and would further increase the length of the statusline.

This doesn't reveal any additional information to the player because
the duration is always fixed, it just makes it easier to keep track of
when it ends.