Shard from Morgoth's Crown inert alt form
The Shard from Morgoth's Crown can now be invoked with no timeout to
turn it into a ring of nothing (new non-randomly-generated ring with a
fixed description of "boring ring") and back again.  The base type's
name (if known) or description (if unknown, to not reveal the identity
of the base type) is also shown in inventory, like for the other
shapeshifting artifacts.  This does mean that the hit message when in
inert form will be "the corrupted boring", but that bug already
existed ("the corrupted twisted" if it's a twisted ring), so this just
doesn't fix that.