Add engulf/grab attacks to avoidable attacks
Similarly to passives/mspells, only shown if you have a grab attack (in current attack chain or wearing grappler's grasp) or an engulf attack (only attack chain). Engulf attacks are easy to avoid, grabs less so.

Normal grab attacks fall back to the case as if the target was also a grabber, which for old nethack sanity code assumes can't be grappled (I guess to avoid you getting stuck to somebody stuck to you. Fair enough, that's ugly). This does have slightly different behaviors in different places - for things like hug attacks, this falls back to a physical damage thing. For something like AD_LAVA, it just fails the grapple check mostly silently. There is unfortunately no exhaustive list of "everytime you try to grab something", but I think I got everything in xhity + artifact.c - there might be stragglers though, unfortunately.

Also fixes the incidental bug of #style options not showing up in #ability unless you have a lightsaber - now, the hasfightingforms function exists to check the form status and properly pass that on to the appropriate place. Also tones down the way it's checked - it's a little bitmask now instead of a bunch of ugly booleans. Works better too.

Thanks to Noisy for some of this code, and for pointing out that the boolean uavoid_x can be made into bitfields - I didn't realize that option was enabled, I thought that was only on like weird compiler setups or something.