add message for 'no artiwishes left', adjust candles to not eat wishes if you get nada
adds a message for "no more artifact wishes left" saying that "alone, such artifacts are beyond my power" / "you cannot wish for an artifact right now".

candles of invocation are modified - now, you can actually exit the menu if you invoked and realized you didn't want it, there's an abort option now that just does nothing but exit the menu. in addition, you're no longer charged for a non-wish, for the most part to make it so that somebody who wanted to wish for an artifact but couldn't for X reason can just save it for later. failing that does the same as failing to generate a summon/dlord, where it kicks you back to your options list (and you can now exit!)

lastly, djinni wishes now message when the djinni/their entourage dips (puff of smoke) since I realized it's a little weird otherwise.

the big change that comes with this is that fucking up your wish 5 times no longer gives you a random item. i think this is fine. now, if you fucked it up it just gives you nothing and returns a failed wish - it's up to the thing that made the wish to determine that behavior. for candles and rings, this does NOT consume a charge. for lamps and i think wands (didn't test them, behavior unmodified afaik) it does. this is indeed a change from long-lasting vanilla behavior but I don't think it's a bad one. the main reason I did this was because there's been a couple times where people tried to wish for X artifact only to get cucked, either because it was already spawned or because they couldn't (mercy rule, didn't realize nowish, etc.) so now they are... less penalized? I think flavor wise leaving lamps & wands 'penalized' makes sense, the summmoning ritual one is iffy (though fixable if you just remove the exit option, probably) and then the ring makes sense and was already half existing i think.

note that fucking up your wish 4 times and wishing correctly on the fifth is still allowed, and you can exit the wish canonically with 'nil'/etc. - or just type gibberish. but wishing for 'nothing' counts as a non-consume, the only consumed wish is a success.