Adjust acurrstr back to weird version
While it would make sense to just chop, in practice the player actually experiences the fractional regions of strength much more than the regions from 19-25, so this function allows some variation in effect in the fractional region at the expense of the 19-25 region.

Another option would be to just make 25 str == 30 in this function, and give the extra 5 points to the fractional region.