Add dbon + information about what causes it to invent lookup
Now, you can actually see both the stats your weapons scale with and your effective dbon. Strength IS dependent on the context of the weapon at the minute, for things like katanas etc.

Refactors the dbon function to use a new atr_dbon function that tells you the scaling, basically, and externs that for use in invent lookup. Also makes OPROP_ELFLW wizwishable for testing purposes.