Edibility, cannibalism, and same-race sacrifice prompt changes
Make them not violate the InterHack principle.  Also, change the
non-incantifer eating prompt from "Stop eating?" to "Continue
eating?".  This makes it consistent with the incantifier eating prompt
and NetHack 3.6.

-The same-race sacrifice confirmation prompt is now always shown,
 since the player always knows whether or not a corpse they are
 sacrificing is the same race as them.
-The cannibalism confirmation prompt is now always shown for
 corpses (corpses are always known).  It is shown for tins only if the
 tin is known or are you not hallucinating.  It is shown for blood
 potions only if the potion is known or will become known if you choose
 to drink it.  For both tins and blood potions, the old
 wisdom/alignment record-based check is used to maybe give the prompt
 anyway even if you don't know the contents.
-Edibility prompts are now shown if they can be derived from
 information that the player knows.  This is mostly based on NetHack4's
 implementation, with several changes:
--You get edibility prompts for tins even without food detection,
  provided that you are not hallucinating (since you are told the
  contents of the tin).
--The overeating prompt was not moved to edibility_prompts().

The cannibalism madness still has some effects, but those that were
exclusively interface screws were removed.  It probably needs a
replacement.

This is a minor balance change: you are now warned about corpses that
might be tainted even without food detection, as in NetHack4.  Food
detection allows you to detect this more precisely.  I think this is a
good thing: previously, you could kill a gnome zombie that you thought
was a gnome due to delusions, or kill an unnamed gnome zombie and
gnome on the same square and not know which corpse is tainted, then
eat the tainted corpse and die because you're a madperson with no
unicorn horn and an angry god.  This has happened to me before.

Everything else is an interface improvement and not a balance change.