curses: Make !classic_status work with !hitpointbar
I don't know why you would want to do this, but now it actually shows
HP instead of just showing "---".

Also replaces a few more uses of sprintf and strcpy with snprintf and
strncpy in cursstat.c.  Some of these cause -Wformat-truncation
warnings, which are silenced with a GCC-specific #pragma, but
truncation is preferable to a buffer overflow.