Add classic_colors option for curses to use "classic" nethack colors
This adds an option to ignore terminal color settings and instead default to 'normal' colors as far as nethack is concerned.

This mostly includes specific definitions for yellow (as brown) and bright red (as orange) which are decidedly non-standard for 16 color terminals, and some specific ideas of 'bright' colors as saturated rather than pastel - so cyan is teal and bright cyan is cyan, for example.

This should cleanup & restore fine on terminals that read 16 colors into curses, but has the same extant bug of not resetting the 8 bright colors for some terminals that only read colors 0-7 into curses. There is no clean solution for this, curses does not support one. Noisy suggested a potential solution of adding a RESETCOLOR option to set which color to restore state to, which is probably the best option. I'll probably implement that later.