Bugfix: Don't override player's configured ghost symbol
assign_graphics() was setting the ghost symbol to the lit room symbol,
so it looked consistent with the floor if you are using a symset with
a fancy floor symbol (like UTF8graphics).  This was overriding the
player's configured ghost symbol.  Instead, only set it if it's
currently set to the default.