Fix the convict and madman hatemasks not being persistently changed.
Drow healers are also cool with fey and elves (but not orcs).

The adjustment was done once when the game was initialized, but in a way that did not persist between save-loads. The fix is to move the code to the role touch up area (where it will be repeated on every load).