Add a command to make pets drop everything they're carrying.
I'm unsure if this is actually needed, but I coded it last month so I should probably push it.

Without this command you're limited to taking one item at a time.

A better solution would be the ability to treat pets as containers, but THAT's not happening for this version.