Fix compilation on BSD and revert to old locale-based check for UTF8
FreeBSD and OpenBSD both define fileno as a macro and require
including sys/stat.h for S_* file mode macros.

Demo and riker both reported issues with the new UTF-8 check from
UnNetHack (it doesn't work on the WSL terminal).  Apparently it works
fine in UnNetHack.  I have no way of testing this and don't know how
to fix it properly, so just revert it for now.