Refactor to reduce impact of getattk()
Was using around 40% of CPU time according to gprof, mostly due to very inefficient use in the weapon selection code.

Refactor to reduce the number of calls to getattack(), primarily by moving mon_attacktype() call out of oslelect into the calling code.

Seems to have reduced impact from >40% of CPU time to around 6% of CPU time.

Classed as a major issue because the inefficiency is causing hardfought to kill the process while traveling etc.